+++ app/src/main/java/com/forrestguice/suntimeswidget/calculator/sunrisesunset_java/SunriseSunsetSuntimesCalculator.java
@@ -18,7 +18,6 @@
 
 package com.forrestguice.suntimeswidget.calculator.sunrisesunset_java;
 
-import android.content.Context;
 import android.util.Log;
 
 import com.forrestguice.suntimeswidget.R;