+++ app/src/main/java/com/forrestguice/suntimeswidget/getfix/GetFixDatabaseAdapter.java
@@ -77,7 +77,7 @@ public class GetFixDatabaseAdapter
     /**
      * Open the database
      * @return a reference the (now open) database adapter
-     * @throws SQLException
+     * @throws SQLException if failed to open
      */
     public GetFixDatabaseAdapter open() throws SQLException
     {
@@ -129,10 +129,11 @@ public class GetFixDatabaseAdapter
      * Get a place from the database
      * @param row the rowID to get
      * @return a Cursor into the database
-     * @throws SQLException
+     * @throws SQLException if query failed
      */
     public Cursor getPlace(long row) throws SQLException
     {
+        @SuppressWarnings("UnnecessaryLocalVariable")
         String[] QUERY = QUERY_PLACES_FULLENTRY;
         Cursor cursor = database.query( true, TABLE_PLACES, QUERY,
                                         KEY_ROWID + "=" + row, null,