+++ app/src/main/java/com/forrestguice/suntimeswidget/layouts/SuntimesLayout_1x1_0.java
@@ -19,6 +19,9 @@
 package com.forrestguice.suntimeswidget.layouts;
 
 import android.content.Context;
+import android.os.Build;
+import android.util.Log;
+import android.util.TypedValue;
 import android.widget.RemoteViews;
 
 import com.forrestguice.suntimeswidget.R;
@@ -68,15 +71,25 @@ public class SuntimesLayout_1x1_0 extends SuntimesLayout
     {
         super.themeViews(context, views, theme);
 
-        // theme sunrise text
         int sunriseColor = theme.getSunriseTextColor();
         int suffixColor = theme.getTimeSuffixColor();
         views.setTextColor(R.id.text_time_sunrise_suffix, suffixColor);
         views.setTextColor(R.id.text_time_sunrise, sunriseColor);
 
-        // theme sunset text
         int sunsetColor = theme.getSunsetTextColor();
         views.setTextColor(R.id.text_time_sunset_suffix, suffixColor);
         views.setTextColor(R.id.text_time_sunset, sunsetColor);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN)
+        {
+            float timeSize = theme.getTimeSizeSp();
+            float suffSize = theme.getTimeSuffixSizeSp();
+
+            views.setTextViewTextSize(R.id.text_time_sunrise_suffix, TypedValue.COMPLEX_UNIT_SP, suffSize);
+            views.setTextViewTextSize(R.id.text_time_sunrise, TypedValue.COMPLEX_UNIT_SP, timeSize);
+
+            views.setTextViewTextSize(R.id.text_time_sunset, TypedValue.COMPLEX_UNIT_SP, timeSize);
+            views.setTextViewTextSize(R.id.text_time_sunset_suffix, TypedValue.COMPLEX_UNIT_SP, suffSize);
+        }
     }
 }