+++ app/src/main/java/com/forrestguice/suntimeswidget/layouts/SuntimesLayout_1x1_1.java
@@ -19,6 +19,8 @@
 package com.forrestguice.suntimeswidget.layouts;
 
 import android.content.Context;
+import android.os.Build;
+import android.util.TypedValue;
 import android.widget.RemoteViews;
 
 import com.forrestguice.suntimeswidget.R;
@@ -64,10 +66,14 @@ public class SuntimesLayout_1x1_1 extends SuntimesLayout
     {
         super.themeViews(context, views, theme);
 
-        // theme sunrise text
         int sunriseColor = theme.getSunriseTextColor();
         int suffixColor = theme.getTimeSuffixColor();
         views.setTextColor(R.id.text_time_sunrise_suffix, suffixColor);
         views.setTextColor(R.id.text_time_sunrise, sunriseColor);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN)
+        {
+            views.setTextViewTextSize(R.id.text_time_sunrise_suffix, TypedValue.COMPLEX_UNIT_SP, theme.getTimeSuffixSizeSp());
+            views.setTextViewTextSize(R.id.text_time_sunrise, TypedValue.COMPLEX_UNIT_SP, theme.getTimeSizeSp());
+        }
     }
 }